Fixed testing of the request context in RequestListener.
Checking `$this->deviceView->isNotMobileView()` does not check if we're in the
request context, it checks if the `mobile_detect.device_view` service was
instantiated when there was already a request, because it sets VIEW_NOT_MOBILE
in the constructor.

Anyway, we’re handling a `kernel.request` event, so we’re always in the request
context at this point.