Fixed request injection in DeviceView.
Getting the request from the Service Container in the constructor and storing
it into a property means that the request will be wrong in subrequests, and
will always be null if the service happens to be instantiated before there is
a request. Therefore, the request is now passed using setter injection, and
the processing that was done in the constructor is now done in the setter.